% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.half.2arm.R
\name{add.half.2arm}
\alias{add.half.2arm}
\title{Transform values (events and no events) and add 0.5 in the no zero cell.}
\usage{
add.half.2arm(tmp)
}
\arguments{
\item{tmp}{is the data frame with number of events and number of total sample size}
}
\value{
a matrix with events and failure data with extra 0.5 if needed.
}
\description{
The function creates events and failure type data and add 0.5 for data including only 2arm studies.
}
\keyword{internal}
