% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont.trans.2arm.R
\name{cont.trans.2arm}
\alias{cont.trans.2arm}
\title{Transform values into LOR values}
\usage{
cont.trans.2arm(tmp, NS = NS)
}
\arguments{
\item{tmp}{is the data file}

\item{NS}{is total number of studies}
}
\value{
a matrix with log odds ratios
}
\description{
This function creates events and failure type data
}
\keyword{internal}
