% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont.trans.R
\name{cont.trans}
\alias{cont.trans}
\title{Transform values and add 0.5}
\usage{
cont.trans(tmp, NS = NS, N2h = N2h)
}
\arguments{
\item{tmp}{is the data file}

\item{NS}{is total number of studies}

\item{N2h}{is number of studies with only 2 arms}
}
\value{
a matrix with log odds ratios
}
\description{
This function create events and failure type data
}
\keyword{internal}
