% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.y.2arm.R
\name{gen.y.2arm}
\alias{gen.y.2arm}
\title{Generate y for bugs input}
\usage{
gen.y.2arm(yr, yd, NS = NS)
}
\arguments{
\item{yr}{for outcome 1}

\item{yd}{for outcome 2}

\item{NS}{is total number of studies}
}
\value{
A vector of combined y as a bugs input
}
\description{
The function generates y for datasets only with 2 arm studies.
}
\keyword{internal}
