% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.y.R
\name{gen.y}
\alias{gen.y}
\title{Generate y for bugs input}
\usage{
gen.y(yr, yd, NS = NS, N2h = N2h)
}
\arguments{
\item{yr}{for outcome 1}

\item{yd}{for outcome 2}

\item{NS}{is total number of studies}

\item{N2h}{is number of studies with only 2 arms.}
}
\value{
A vector of combined y as a bugs input
}
\description{
The function generates y for bivariate anlaysis based on NS and N2h
}
\keyword{internal}
