% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnma.model.R
\name{mnma.model}
\alias{mnma.model}
\title{Create a data object for multivariate and univariate network meta-analysis (NMA)}
\usage{
mnma.model(outcome1, outcome2, TX, noa = "3arm", ref = 1,
  parameters = "simple")
}
\arguments{
\item{outcome1}{A data frame for outcome 1. First 2 (or 3) columns contain the information about number of events for each arm and the rest 2 (or 3) columns contain the sample size for each arm. Number of columns ranges from 2 to 3, depending on the number of maximum arms in the trial data.}

\item{outcome2}{A data frame for outcome 2, which has the same dimensions as outcome1.}

\item{TX}{A data frame for tratments, which has maximum 3 treatment arms in 3 columns. The number of rows should coincide with outcome1 and outcome2. Each element needs to be numeric.}

\item{noa}{Maximum number of arms in the trial studies, which is either '2arm' or '3arm'. Note that number of arm is different from number of total treatments in the network.}

\item{ref}{Reference treatment in the analysis in a numeric class.}

\item{parameters}{List of parameters to infer in the analysis. They are either 'simple' or 'details' depending on comparisons.}
}
\value{
An object of class 'mnma.model' that contains data information and parameter lists to estimate for multivariate and univariate NMA.
}
\description{
The function generates a data object for BUGS operation, using outcome 1, outcome 2, and treatments inputs
}
\details{
A 'simple' list for parameters includes log odds ratio (comparison) parameters between reference and all other treatments. A 'detail' list for parameters includes log odds ratio parameters between all treatments. Both cases include a correlation parameter and heterogeneity parameters for two outcomes.
}
\examples{
data(acute)
# Split the dataset to create a mnma data object
outcome1 <- acute[,c(1:6)]     # outcome1
outcome2 <- acute[,c(7:12)]    # outcome2
TX    <- acute[,13:15]
# Convert them into log odds ratio inputs
data <- mnma.model(outcome1,outcome2,TX, ref=1, parameters="simple")
data

}
