% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnma.run.R
\name{mnma.run}
\alias{mnma.run}
\title{Run multivariate network meta-analysis (NMA) anlaysis}
\usage{
mnma.run(input, inits = NULL, n.iter = 5000, n.burnin = 2000,
  n.chains = 1, n.thin = 1, debug = FALSE, codaPkg = FALSE)
}
\arguments{
\item{input}{\code{mnma.model} class data, which can be generated using \code{mnma.model} function}

\item{inits}{a list of initial values for \bold{OpenBUGS} model with \code{n.chains} elements; If \code{inits=NULL} (default), initial values are randomly generated by \bold{OpenBUGS}.}

\item{n.iter}{number of total iterations per chain (default: 5000)}

\item{n.burnin}{length of burnin per chain (default: 2000)}

\item{n.chains}{number of Markov chains (default: 1)}

\item{n.thin}{thining rate (default: 1)}

\item{debug}{if FALSE (default), \strong{OpenBUGS} is closed automatically after running the \code{mnma.run} script. If TRUE, \strong{OpenBUGS} remains open with additional information.}

\item{codaPkg}{if FALSE (default), a \code{bugs} object is returned. Otherwise, file names of \strong{OpenBUGS} output are returned with TRUE.}
}
\value{
\code{mnma.result} class \code{bugs} result. See \code{\link[R2OpenBUGS]{bugs}} for details.
}
\description{
The function conducts bivariate NMA using a \code{mnma.model} object as an input
}
\examples{
\dontrun{
# Run after data transformation using mnma.model function.
res <- mnma.run(data, n.iter=3000, n.burnin=1000)
names(res)
res$summary  # provides mean, standard deviation, and percentile information about log odds ratio
}


}
\references{
Efthimiou et al. (2015) Biostatistics 16(1):84-97
(\href{https://www.ncbi.nlm.nih.gov/pubmed/24992934}{PubMed})
}
\seealso{
\code{\link[R2OpenBUGS]{bugs}}
}
