% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mul.model.R
\name{mul.model}
\alias{mul.model}
\title{Create a data object for multivariate network meta-analysis (MNMA)}
\usage{
mul.model(data1, data2, TX, noa = "3arm", ref = 1, parameters = "simple")
}
\arguments{
\item{data1}{A data frame for outcome 1. First 2 (or 3) columns contain the information about number of events for each arm and the rest 2 (or 3) colums contain the sample size for each arm. Number of columns is from 2 to 3, depending on the number of maximum arms in the data.}

\item{data2}{A data frame for outcome 2, which has the same number of rows and columns.}

\item{TX}{A data frame for tratments, which has maximum 3 treatments in 3 columns. The number of rows should coincide with data1 and data2.}

\item{noa}{Maximum number of arms in the trial studies.}

\item{ref}{reference treatment numbers in the analysis}

\item{parameters}{'simple' or 'details' list of parameters.}
}
\value{
mnma.model object for mnma.run operation
}
\description{
The function creates a data frame for bugs operation, using outcome 1, outcome 2, and treatments inputs
}
\examples{
data(acute)
# Split the dataset
data1 <- acute[,c(1:6)]     # outcome1
data2 <- acute[,c(7:12)]    # outcome2
TX    <- acute[,13:15]
# Convert them into log odds ratio inputs
data <- mul.model(data1,data2,TX, ref=1, noa="3arm", parameters="simple")
data

}
\keyword{internal}
