% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortbyna.2arm.R
\name{sortbyna.2arm}
\alias{sortbyna.2arm}
\title{Sort data by number of arms for trial data only with 2 arm studies.}
\usage{
sortbyna.2arm(out1, out2, TX)
}
\arguments{
\item{out1}{is the dataframe for outcome 1}

\item{out2}{is the dataframe for outcome 2}

\item{TX}{is the dataframe for treatment}
}
\value{
combined sorted data.frame with additinonal column for number of arms and number of outcome
}
\description{
The function sort existing data and create a sorted dataset depending on the number of arms
}
\keyword{internal}
