% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortbyna.R
\name{sortbyna}
\alias{sortbyna}
\title{Sort data by number of arms in the trial}
\usage{
sortbyna(out1, out2, TX)
}
\arguments{
\item{out1}{is the dataframe for outcome 1}

\item{out2}{is the dataframe for outcome 2}

\item{TX}{is the dataframe for treatment}
}
\value{
combined sorted data.frame with additinonal column for number of arms and number of outcome
}
\description{
The function sort existing data and create a sorted dataset depending on the number of arms
}
\keyword{internal}
