% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx.check.R
\name{tx.check}
\alias{tx.check}
\title{Check the treatment data}
\usage{
tx.check(TX)
}
\arguments{
\item{TX}{is the data frame contains treatment 1,2, or 3 and # of columns have to be half of that of data1.}
}
\value{
It returns the # of treatment.
}
\description{
Check the any missing numbering in the treatments and shows the warning sign if needed
}
\keyword{internal}
