% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnma.run.R
\name{mnma.run}
\alias{mnma.run}
\title{Run multivariate NMA anlaysis}
\usage{
mnma.run(input, inits = NULL, n.iter = 5000, n.burnin = 2000,
  n.chains = 1, n.thin = 1, debug = FALSE, codaPkg = FALSE)
}
\arguments{
\item{input}{mnma.model class data, which can be generated using mnma.model function}

\item{inits}{a list with n.chains element (See \code{\link{bugs}})}

\item{n.iter}{number of total iterations per chain (default: 5000)}

\item{n.burnin}{length of burnin per chain (default: 2000)}

\item{n.chains}{number of Markov chains (default: 1)}

\item{n.thin}{Thining (default: 1)}

\item{debug}{Openbugs is closed automatically when finishing running the bugs script with the FALSE setting (default).}

\item{codaPkg}{if FALSE (default) a bugs object is returned. (See and check \code{\link{bugs}})}
}
\value{
mnma.result class bugs result. See \code{\link{bugs}} for details.
}
\description{
The function conducts bivariate NMA using a mnma.model object as an input
}
\examples{
\dontrun{
# Run after data transformation using mnma.model function.
res <- mnma.run(data, n.iter=3000, n.burnin=1000)
res$summary
}


}
\references{
Efthimiou et al. (2015) Biostatistics 16(1):84-97
(\href{https://www.ncbi.nlm.nih.gov/pubmed/24992934}{PubMed})
}
\seealso{
\code{\link{bugs}}
}
